lab8/requirements.txt
Pillow==10.1.0
matplotlib==3.8.2
numpy==1.26.2
